+++ src/bundle/Resources/schema/legacy_matrix.xsd
@@ -4,11 +4,6 @@
   ~ @license For full copyright and license information view LICENSE file distributed with this source code.
   -->
 
-<!--
-  ~ @copyright Copyright (C) eZ Systems AS. All rights reserved.
-  ~ @license For full copyright and license information view LICENSE file distributed with this source code.
-  -->
-
 <!-- Legacy ezmatrix format:
 
 <?xml version="1.0" encoding="utf-8"?>
